move commands to own files
